+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add a new item template creation window which lists commonly used binding configurations
 - Add the ability to create groups/folders to organize bindings
 - Add a class load option
+- Add a player name-realm load option
 - Automatically update listed specializations, talents, and PvP talents based on the selected class/spec load option
 - Remember the selected binding page (action, targets, conditions) when switching between bindings